# server/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DD Core API",
    "version": "0.1.0",
    "description": "Stable v1 contract for DD instance discovery, service health, account authentication, contacts, and P4 reliable text messaging/sync. P0 call endpoints remain experimental and are intentionally excluded until promoted into /api/v1.",
    "license": {
      "name": "License decision pending",
      "url": "https://dd.invalid/license-pending"
    }
  },
  "servers": [
    {
      "url": "https://{instanceHost}",
      "description": "Self-hosted DD instance",
      "variables": {
        "instanceHost": {
          "default": "dd.invalid",
          "description": "Hostname of the self-hosted instance"
        }
      }
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/me/avatar": {
      "put": {
        "operationId": "updateCurrentUserAvatar",
        "summary": "Upload or replace the current user's avatar",
        "description": "Accepts JPEG, PNG or WebP image bytes up to 2 MiB. SVG and other active formats are rejected.",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "image/jpeg": {"schema": {"type": "string", "format": "binary"}},
            "image/png": {"schema": {"type": "string", "format": "binary"}},
            "image/webp": {"schema": {"type": "string", "format": "binary"}}
          }
        },
        "responses": {
          "200": {"description": "Avatar updated"},
          "400": {"description": "Avatar type or bytes are invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "413": {"description": "Avatar exceeds 2 MiB", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      },
      "delete": {
        "operationId": "deleteCurrentUserAvatar",
        "summary": "Remove the current user's avatar",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Avatar removed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OperationResultResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
    "/api/v1/avatars/{userId}": {
      "get": {
        "operationId": "getUserAvatar",
        "summary": "Get one active user's avatar",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "userId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {
            "description": "Avatar image bytes",
            "content": {
              "image/jpeg": {"schema": {"type": "string", "format": "binary"}},
              "image/png": {"schema": {"type": "string", "format": "binary"}},
              "image/webp": {"schema": {"type": "string", "format": "binary"}}
            }
          },
          "400": {"description": "User id is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "404": {"description": "Avatar was not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
    "/api/v1/push/endpoints": {
      "get": {
        "operationId": "listPushEndpoints",
        "summary": "List push endpoint metadata for the current authenticated device",
        "description": "Raw provider tokens and capability URLs are deliberately not returned.",
        "tags": ["Push"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Current device push endpoint metadata", "content": {"application/json": {"schema": {"type": "object"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      },
      "post": {
        "operationId": "registerPushEndpoint",
        "summary": "Register or rotate a push endpoint for the current authenticated device",
        "tags": ["Push"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["provider", "endpoint"], "properties": {"provider": {"type": "string", "enum": ["FCM", "APNS", "UNIFIEDPUSH"]}, "endpoint": {"type": "string", "minLength": 1, "maxLength": 4096}, "appId": {"type": "string", "maxLength": 160}, "environment": {"type": "string", "enum": ["PRODUCTION", "SANDBOX"]}}, "additionalProperties": false}}}},
        "responses": {
          "200": {"description": "Endpoint registered", "content": {"application/json": {"schema": {"type": "object"}}}},
          "400": {"description": "Endpoint is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "403": {"description": "Authenticated device is revoked or does not belong to the principal", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "409": {"description": "Endpoint is already bound to another account", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      },
      "put": {
        "operationId": "putPushEndpoint",
        "summary": "Idempotently register or rotate a push endpoint for the current device",
        "tags": ["Push"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["provider", "endpoint"], "properties": {"provider": {"type": "string", "enum": ["FCM", "APNS", "UNIFIEDPUSH"]}, "endpoint": {"type": "string", "minLength": 1, "maxLength": 4096}, "appId": {"type": "string", "maxLength": 160}, "environment": {"type": "string", "enum": ["PRODUCTION", "SANDBOX"]}}, "additionalProperties": false}}}},
        "responses": {
          "200": {"description": "Endpoint registered", "content": {"application/json": {"schema": {"type": "object"}}}},
          "400": {"description": "Endpoint is invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "403": {"description": "Authenticated device is revoked or does not belong to the principal", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "409": {"description": "Endpoint is already bound to another account", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OperationResultData": {
        "type": "object",
        "additionalProperties": true
      },
      "OperationResultResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data", "requestId"],
        "properties": {"data": {"$ref": "#/components/schemas/OperationResultData"}, "requestId": {"$ref": "#/components/schemas/RequestIdValue"}}
      },
...
      "RequestIdValue": {
        "type": "string",
        "pattern": "^req_[A-Za-z0-9]+$"
      },
      "APIError": {
        "type": "object",
        "additionalProperties": true,
        "required": ["code", "message", "requestId"],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestIdValue"
          },
          "details": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        }
      }
    }
  }
}